% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_model.R
\name{SpatialModel}
\alias{SpatialModel}
\title{R6 class for SpatialVectorModel}
\description{
R6 class for SpatialVectorModel

R6 class for SpatialVectorModel
}
\details{
This class is used to fit Bayesian model when the response is a vector with associated coordiantes for each value.
}
\section{Public fields}{

\itemize{
\item \code{name}: name of the model.
\item \code{event_initial}: The time when the event starts.
\item \code{vector_name}: name of the response variable (only used in plots).
\item \code{variables_names}: name of each of the components of vector (only used in plots).
\item \code{confidence_level}: level of the credible interval.
\item \code{n_simul}: number of the iteration for MCMC, half of the iteration will be burned, default 2000.
\item \code{n_chains}: number of chains for MCMC, default 4.
\item \code{n_cores}: number of cores used in MCMC, if NA(default) the number of CPU cores will be used.
\item \code{thin}: thinning used in MCMC, default 1.
\item \code{n_spatial_kernels}: number of spatial kernels to use, default 2.
}
}

\section{Super classes}{
\code{\link[dynamicimpact:BaseImpactModel]{dynamicimpact::BaseImpactModel}} -> \code{\link[dynamicimpact:StanModelVector]{dynamicimpact::StanModelVector}} -> \code{SpatialVectorModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SpatialModel$new()}}
\item \href{#method-fit}{\code{SpatialModel$fit()}}
\item \href{#method-get_unscaled_centroids}{\code{SpatialModel$get_unscaled_centroids()}}
\item \href{#method-plot_map}{\code{SpatialModel$plot_map()}}
\item \href{#method-clone}{\code{SpatialModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dynamicimpact" data-topic="BaseImpactModel" data-id="get_dates_df">}\href{../../dynamicimpact/html/BaseImpactModel.html#method-get_dates_df}{\code{dynamicimpact::BaseImpactModel$get_dates_df()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dynamicimpact" data-topic="BaseImpactModel" data-id="get_end_time">}\href{../../dynamicimpact/html/BaseImpactModel.html#method-get_end_time}{\code{dynamicimpact::BaseImpactModel$get_end_time()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dynamicimpact" data-topic="StanModelVector" data-id="plot">}\href{../../dynamicimpact/html/StanModelVector.html#method-plot}{\code{dynamicimpact::StanModelVector$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dynamicimpact" data-topic="StanModelVector" data-id="plot_aggregate">}\href{../../dynamicimpact/html/StanModelVector.html#method-plot_aggregate}{\code{dynamicimpact::StanModelVector$plot_aggregate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dynamicimpact" data-topic="StanModelVector" data-id="plot_diagnosis">}\href{../../dynamicimpact/html/StanModelVector.html#method-plot_diagnosis}{\code{dynamicimpact::StanModelVector$plot_diagnosis()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dynamicimpact" data-topic="StanModelVector" data-id="plot_individual">}\href{../../dynamicimpact/html/StanModelVector.html#method-plot_individual}{\code{dynamicimpact::StanModelVector$plot_individual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dynamicimpact" data-topic="StanModelVector" data-id="predict">}\href{../../dynamicimpact/html/StanModelVector.html#method-predict}{\code{dynamicimpact::StanModelVector$predict()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialModel$new(
  name = "Model Vector spatial",
  event_initial,
  X_data,
  Y_data,
  vector_name,
  variables_names,
  confidence_level,
  n_simul = 20000,
  n_chains = 4,
  n_cores = NA,
  predefined_cov_matrix = NULL,
  stan_fit = NA_real_,
  n_spatial_kernels = 2,
  thin = 1,
  dates = NULL,
  coordinates
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name:}}{name of the model.}

\item{\code{y_data:}}{response matrix.}

\item{\code{X_data:}}{covariates matrix.}

\item{\code{event_initial:}}{The time when the event starts.}

\item{\code{vector_name:}}{name of the response variable (only used in plots).}

\item{\code{variables_names:}}{name of each of the components of vector (only used in plots).}

\item{\code{confidence_level:}}{level of the credible interval.}

\item{\code{n_simul:}}{number of the iteration for MCMC, half of the iteration will be burned, default 2000.}

\item{\code{n_chains:}}{number of chains for MCMC, default 4.}

\item{\code{n_cores:}}{number of cores used in MCMC, if NA(default) the number of CPU cores will be used.}

\item{\code{thin:}}{thinning used in MCMC, default 1.}

\item{\code{predefined_cov_matrix:}}{If you want used a predefined covariance matrix you can used this variable.}

\item{\code{predefined_cov_matrix_type:}}{The covariance matrix type, if = mcmc  the covariance matrix is is estimated using MCMC, if ml the covariance matrix is estimated used maximun likehood and then "plugin" as constant in MCMC, if = identity the covariance matrix is "plugin"}

\item{\code{`coordinates`:}}{matrix of coordinates, must be in a planar projection.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \code{SpatialModel} object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialModel$fit()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Fit the model using MCMC.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_unscaled_centroids"></a>}}
\if{latex}{\out{\hypertarget{method-get_unscaled_centroids}{}}}
\subsection{Method \code{get_unscaled_centroids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialModel$get_unscaled_centroids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_map"></a>}}
\if{latex}{\out{\hypertarget{method-plot_map}{}}}
\subsection{Method \code{plot_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialModel$plot_map(base_shape = NA, coord_system)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpatialModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
