% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_vector}
\alias{X_vector}
\title{Simulated input data.}
\format{
a matrix of 122 rows and 2 columns:
}
\usage{
X_vector
}
\description{
A matrix used as example in the package.
}
\keyword{datasets}
