% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_vector}
\alias{Y_vector}
\title{Simulated response data.}
\format{
a matrix of 122 rows and 2 columns:
}
\usage{
Y_vector
}
\description{
A matrix used as example in the package.
}
\keyword{datasets}
