% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_model.R
\name{StanModelVector}
\alias{StanModelVector}
\title{R6 class for StanModelVector}
\description{
R6 class for StanModelVector

R6 class for StanModelVector
}
\details{
This class is used to fit Bayesian model when the response is a vector.
}
\section{Public fields}{

\itemize{
\item \code{name}: name of the model.
\item \code{event_initial}: The time when the event starts.
\item \code{vector_name}: name of the response variable (only used in plots).
\item \code{variables_names}: name of each of the components of vector (only used in plots).
\item \code{confidence_level}: level of the credible interval.
\item \code{n_simul}: number of the iteration for MCMC, half of the iteration will be burned, default 2000.
\item \code{n_chains}: number of chains for MCMC, default 4.
\item \code{n_cores}: number of cores used in MCMC, if NA(default) the number of CPU cores will be used.
\item \code{thin}: thinning used in MCMC, default 1.
}
}

\section{Super class}{
\code{\link[dynamicimpact:BaseImpactModel]{dynamicimpact::BaseImpactModel}} -> \code{StanModelVector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanModelVector-new}{\code{StanModelVector$new()}}
\item \href{#method-StanModelVector-fit}{\code{StanModelVector$fit()}}
\item \href{#method-StanModelVector-predict}{\code{StanModelVector$predict()}}
\item \href{#method-StanModelVector-plot_individual}{\code{StanModelVector$plot_individual()}}
\item \href{#method-StanModelVector-plot_aggregate}{\code{StanModelVector$plot_aggregate()}}
\item \href{#method-StanModelVector-plot}{\code{StanModelVector$plot()}}
\item \href{#method-StanModelVector-plot_diagnosis}{\code{StanModelVector$plot_diagnosis()}}
\item \href{#method-StanModelVector-summary}{\code{StanModelVector$summary()}}
\item \href{#method-StanModelVector-clone}{\code{StanModelVector$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dynamicimpact" data-topic="BaseImpactModel" data-id="get_dates_df"><a href='../../dynamicimpact/html/BaseImpactModel.html#method-BaseImpactModel-get_dates_df'><code>dynamicimpact::BaseImpactModel$get_dates_df()</code></a></li>
<li><span class="pkg-link" data-pkg="dynamicimpact" data-topic="BaseImpactModel" data-id="get_end_time"><a href='../../dynamicimpact/html/BaseImpactModel.html#method-BaseImpactModel-get_end_time'><code>dynamicimpact::BaseImpactModel$get_end_time()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$new(
  name = "Model Vector",
  event_initial,
  X_data,
  Y_data,
  vector_name = "vector model",
  variables_names,
  confidence_level,
  n_simul = 2000,
  n_chains = 4,
  n_cores = NA,
  predefined_cov_matrix_type = c("mcmc", "ml", "identity"),
  predefined_cov_matrix = NULL,
  stan_fit = NA_real_,
  thin = 1,
  log_x = FALSE,
  log_y = FALSE,
  dates = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name:}}{name of the model.}

\item{\code{Y_data:}}{response matrix.}

\item{\code{X_data:}}{covariates matrix.}

\item{\code{event_initial:}}{The time when the event starts.}

\item{\code{vector_name:}}{name of the response variable (only used in plots).}

\item{\code{variables_names:}}{name of each of the components of vector (only used in plots).}

\item{\code{confidence_level:}}{level of the credible interval.}

\item{\code{n_simul:}}{number of the iteration for MCMC, half of the iteration will be burned, default 2000.}

\item{\code{n_chains:}}{number of chains for MCMC, default 4.}

\item{\code{n_cores:}}{number of cores used in MCMC, if NA(default) the number of CPU cores will be used.}

\item{\code{thin:}}{thinning used in MCMC, default 1.}

\item{\code{predefined_cov_matrix:}}{If you want used a predefined covariance matrix you can used this variable.}

\item{\code{predefined_cov_matrix_type:}}{The covariance matrix type, if = mcmc  the covariance matrix is is estimated using MCMC, if ml the covariance matrix is estimated used maximun likehood and then "plugin" as constant in MCMC, if = identity the covariance matrix is "plugin"}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \code{StanModelVector} object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-fit"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$fit()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Fit the model using MCMC.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-predict"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$predict(event_initial = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_initial}}{time when event of interest start, if NULL the event_initial set in the initialization is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make a prediction using MCMC.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-plot_individual"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-plot_individual}{}}}
\subsection{Method \code{plot_individual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$plot_individual(plot_variable, event_initial = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_variable}}{name of the variable to plot, the name must be in the vector variables_names.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
plot a single variable.
}

\subsection{Returns}{
a ggplot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-plot_aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-plot_aggregate}{}}}
\subsection{Method \code{plot_aggregate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$plot_aggregate(event_initial = NULL)}\if{html}{\out{</div>}}
}

\subsection{Details}{
plot an aggregate of all variables.
}

\subsection{Returns}{
a ggplot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-plot"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$plot()}\if{html}{\out{</div>}}
}

\subsection{Details}{
same as plot_aggregate.
}

\subsection{Returns}{
a ggplot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-plot_diagnosis"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-plot_diagnosis}{}}}
\subsection{Method \code{plot_diagnosis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$plot_diagnosis(
  parameter_name = "theta_vec",
  measure_name = "Rhat",
  plot_type = c("interval", "boxplot"),
  event_initial = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter_name}}{name of the parameter of the model to plot, if parameter_name=NA_character_ and plot_type = boxplot all of the parameters are ploted.}

\item{\code{measure_name}}{name of the measure plot must one of: "Rhat", "n_eff", "mean".}

\item{\code{plot_type}}{The possible values are interval for a confidence interval or boxplot.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
plot a diagnosis measure.
}

\subsection{Returns}{
a ggplot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-summary"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$summary(dates_list, confidence_level = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelVector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelVector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelVector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
